Consolidate Edit component
